pyproject.toml
[build-system]
requires = ["setuptools>=62.3"]
build-backend = "setuptools.build_meta"

[project]
name        = "hassil"
version     = "3.1.1"
license     = {text = "Apache-2.0"}
description = "The Home Assistant Intent Language parser"
readme      = "README.md"
authors     = [
    {name = "The Home Assistant Authors", email = "hello@home-assistant.io"}
]
keywords    = ["home", "assistant", "intent", "recognition"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Text Processing :: Linguistic",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.9.0"
dependencies = [
    "PyYAML>=6.0,<7",
    "unicode-rbnf>=2.3,<3"
]

[project.optional-dependencies]
dev = [
    "black==24.8.0",
    "flake8==7.2.0",
    "mypy==1.14.0",
    "pylint==3.2.7",
    "pytest==8.3.5",
    "tox==4.26.0",
    "build==1.2.2",
    "home-assistant-intents==2025.7.30",
]

[project.urls]
"Source Code" = "http://github.com/OHF-Voice/hassil"

[tool.setuptools]
platforms = ["any"]
zip-safe  = true
include-package-data = true

[tool.setuptools.packages.find]
include = ["hassil"]
exclude = ["tests", "tests.*"]

[project.scripts]
hassil = "hassil.__main__:main"
